requirements.txt
click == 7.0
pandas == 0.25.0
pandas-schema == 0.3.4
pyarrow == 0.14.1
xlrd == 1.2.0
PyYAML == 5.1.2
